Create get user listings API Route

+++ server/controllers/user.controller.js
@@ -1,13 +1,28 @@
-import { errorHandler } from "../utils/error.js"
 import bcryptjs from 'bcryptjs';
-import User from "../models/user.model.js"
+import User from '../models/user.model.js';
+import { errorHandler } from '../utils/error.js';
+import Listing from '../models/listing.model.js';
 export const test = (req, res) => {
     res.json({
         message: "user route and controller"
     })
 }
+export const getUserListings = async (req, res, next) => {
+    //console.log("req.user._id:", req.user.id);
+    //console.log("req.params.id:", req.params.id);
+    if (req.user._id === req.params.id) {
+        try {
+            const listings = await Listing.find({ userRef: req.params.id })
+            res.status(200).json(listings)
+        } catch (error) {
+            next(error)
+        }
+    } else {
+        return next(errorHandler(401, 'You can only view your own listings!'));
+    }
+}
 export const updateuser = async (req, res, next) => {
-    //if (req.user.id !== req.params.id) return next(errorHandler(401, 'You are not authorized to do that !!'));
+    if (req.user._id !== req.params.id) return next(errorHandler(401, 'You are not authorized to do that !!'));
 
     try {
         if (req.body.password) {
@@ -28,7 +43,7 @@ export const updateuser = async (req, res, next) => {
     }
 };
 export const deleteuser = async (req, res, next) => {
-    //if (req.user.id !== req.params.id) return next(errorHandler(401, 'You can only delete your account  !!'));
+    if (req.user._id !== req.params.id) return next(errorHandler(401, 'You can only delete your account  !!'));
     try {
         await User.findByIdAndDelete(req.params.id)
         res.clearCookie('access_token')
